moved blog card to it's page dir
